int not long

+++ smarts/trunk/workspace/Tomcat-Common/src/smartshift/common/hibernate/dao/tasks/RowCountTask.java
@@ -15,7 +15,7 @@ import smartshift.common.util.log4j.SmartLogger;
  * 
  * task to get row count based on criterion
  */
-public class RowCountTask<T> extends BaseHibernateTask<T, Integer>{
+public class RowCountTask<T> extends BaseHibernateTask<T, Long>{
     private static final SmartLogger logger = new SmartLogger(RowCountTask.class);
     
     private final Criterion[] _criterions;
@@ -34,14 +34,14 @@ public class RowCountTask<T> extends BaseHibernateTask<T, Integer>{
      * @see smartshift.common.hibernate.dao.tasks.BaseHibernateTask#executeWithSession(org.hibernate.Session)
      */
     @Override
-    public Integer executeWithSession(Session session) throws HibernateException {
+    public Long executeWithSession(Session session) throws HibernateException {
         logger.debug("Enter. Criterions: " + _criterions.length);
         Criteria criteria = session.createCriteria(getDAO().getModelClass());
         for(Criterion criterion : _criterions) {
             logger.trace("Adding criteria: " + criteria.toString());
             criteria.add(criterion);
         }
-        Integer rowCount = (Integer) criteria.setProjection(Projections.rowCount()).uniqueResult();
+        Long rowCount = (Long) criteria.setProjection(Projections.rowCount()).uniqueResult();
         logger.debug("Exit. Got: " + rowCount);
         return rowCount;
     }
